+++ src/main/java/com/clientes/cuentas/demo/application/usecase/GetCustomersWithHigherAmountUseCase.java
@@ -2,6 +2,7 @@ package com.clientes.cuentas.demo.application.usecase;
 
 import com.clientes.cuentas.demo.domain.model.Customer;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -15,5 +16,5 @@ public interface GetCustomersWithHigherAmountUseCase {
    * @param amount the minimum total balance that the sum of all bank accounts associated with a customer must exceed
    * @return a list of {@link Customer} whose aggregated account balance is greater than the specified amount
    */
-  List<Customer> execute(Double amount);
+  List<Customer> execute(BigDecimal amount);
 }
