+++ src/main/java/com/clientes/cuentas/demo/infrastructure/persistence/repository/JpaCustomerRepository.java
@@ -7,6 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
@@ -82,7 +83,7 @@ public interface JpaCustomerRepository extends JpaRepository<CustomerEntity, Lon
           GROUP BY c
           HAVING SUM(ba.total) > :amount
           """)
-  List<CustomerEntity> getCustomersWithHigherAmount(@Param("amount") Double amount);
+  List<CustomerEntity> getCustomersWithHigherAmount(@Param("amount") BigDecimal amount);
 
   Optional<CustomerEntity> findByDni(String dni);
 }
